Add basic files and structure for Falcon app
This commit is a rather large one. It contains
everything necessary to run a test Falcon application
that spits out "Hello World".

This patch also includes a setup.py as well as a Dockerfile
for running locally and through a container.

Finally, this patch adds tox for testing. Currently,
the test dir is empty, but it will run pep8 and will
be there for eventual unit tests.