Add redis support and data combination
This patch has a decent amount of changes in it.
First of all, it hooks with a redis db to gather price reqs.

Secondly, it combines external HTTP and db
requests into one dictionary to be delivered as json
by Falcon.

Finally, the readme is updated to reflect host network
for container communication.