Uses ID that is passed in with REST call
You can now call a GET request and pass in
ID. If the ID is found in the "external api"
(weblink), it will get all of the necessary
values from the external api as well as the
database.

There is no error handling currently and that
is something I'd like to add if there is time.